requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
backports.tarfile==1.2.0
cachetools==5.5.2
campay==1.0.9
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
cssselect==1.3.0
cssselect2==0.8.0
cssutils==2.11.1
defusedxml==0.7.1
Django==5.2
django-allauth==65.7.0
django-ratelimit==4.1.0
Faker==37.1.0
html5lib==1.1
idna==3.10
importlib_metadata==8.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
keyring==25.6.0
lxml==5.3.2
more-itertools==10.6.0
oauthlib==3.2.2
oscrypto==1.3.0
pillow==11.1.0
premailer==3.10.0
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.10.1
pypdf==5.4.0
python-bidi==0.6.6
python-dotenv==1.1.0
python-social-auth==0.3.6
python3-openid==3.2.0
PyYAML==6.0.2
qrcode==8.1
ratelimit==2.2.1
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
SecretStorage==3.3.3
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
xhtml2pdf==0.2.17
yagmail==0.15.293
zipp==3.21.0
